cleaned up some of the client side redirect
