+++ shopify.app.toml
@@ -2,7 +2,7 @@
 
 client_id = "add7504331733e92252623482056ef16"
 name = "app next template"
-application_url = "https://trim-fishing-foo-difference.trycloudflare.com"
+application_url = "https://invision-crowd-cvs-andrea.trycloudflare.com"
 embedded = true
 
 [build]
@@ -16,9 +16,9 @@ embedded_app_direct_api_access = true
 
 [auth]
 redirect_urls = [
-  "https://trim-fishing-foo-difference.trycloudflare.com/auth/callback",
-  "https://trim-fishing-foo-difference.trycloudflare.com/auth/shopify/callback",
-  "https://trim-fishing-foo-difference.trycloudflare.com/api/auth/callback"
+  "https://invision-crowd-cvs-andrea.trycloudflare.com/auth/callback",
+  "https://invision-crowd-cvs-andrea.trycloudflare.com/auth/shopify/callback",
+  "https://invision-crowd-cvs-andrea.trycloudflare.com/api/auth/callback"
 ]
 
 [webhooks]
